move DetourLinksSet to ParallelPathDataProvider
