+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/insertion/ParallelPathDataProvider.java
@@ -20,6 +20,8 @@ package org.matsim.contrib.drt.optimizer.insertion;
 
 import static org.matsim.contrib.drt.optimizer.VehicleData.Entry;
 
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
@@ -28,9 +30,9 @@ import java.util.concurrent.Future;
 
 import javax.inject.Named;
 
+import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.network.Network;
-import org.matsim.contrib.drt.optimizer.insertion.DetourLinksProvider.DetourLinksSet;
 import org.matsim.contrib.drt.passenger.DrtRequest;
 import org.matsim.contrib.drt.run.DrtConfigGroup;
 import org.matsim.contrib.dvrp.path.OneToManyPathSearch;
@@ -45,6 +47,34 @@ import org.matsim.core.router.util.TravelTime;
  * @author michalm
  */
 public class ParallelPathDataProvider implements PrecalculablePathDataProvider, MobsimBeforeCleanupListener {
+
+	private static class DetourLinksSet {
+		final Map<Id<Link>, Link> pickupDetourStartLinks;
+		final Map<Id<Link>, Link> pickupDetourEndLinks;
+		final Map<Id<Link>, Link> dropoffDetourStartLinks;
+		final Map<Id<Link>, Link> dropoffDetourEndLinks;
+
+		public DetourLinksSet(List<InsertionGenerator.Insertion> filteredInsertions) {
+			pickupDetourStartLinks = new HashMap<>();
+			pickupDetourEndLinks = new HashMap<>();
+			dropoffDetourStartLinks = new HashMap<>();
+			dropoffDetourEndLinks = new HashMap<>();
+
+			filteredInsertions.forEach(insertion -> {
+				addLink(pickupDetourStartLinks, insertion.pickup.previousLink);
+				addLink(pickupDetourEndLinks, insertion.pickup.nextLink);
+				addLink(dropoffDetourStartLinks, insertion.dropoff.previousLink);
+				addLink(dropoffDetourEndLinks, insertion.dropoff.nextLink);
+			});
+		}
+
+		private void addLink(Map<Id<Link>, Link> map, Link link) {
+			if (link != null) {
+				map.put(link.getId(), link);
+			}
+		}
+	}
+
 	public static final int MAX_THREADS = 4;
 
 	private final OneToManyPathSearch toPickupPathSearch;
@@ -73,7 +103,7 @@ public class ParallelPathDataProvider implements PrecalculablePathDataProvider, 
 	}
 
 	@Override
-	public void precalculatePathData(DrtRequest drtRequest, DetourLinksSet detourLinksSet) {
+	public void precalculatePathData(DrtRequest drtRequest, List<InsertionGenerator.Insertion> filteredInsertions) {
 		Link pickup = drtRequest.getFromLink();
 		Link dropoff = drtRequest.getToLink();
 
@@ -84,8 +114,11 @@ public class ParallelPathDataProvider implements PrecalculablePathDataProvider, 
 		// with vehicle insertion filtering -- pathsToPickup is the most computationally demanding task, while
 		// pathsFromDropoff is the least demanding one
 
+		DetourLinksSet detourLinksSet = new DetourLinksSet(filteredInsertions);
+
 		// highest computation time (approx. 45% total CPU time)
 		Future<Map<Link, PathData>> pathsToPickupFuture = executorService.submit(() -> {
+			//TODO move extraction of links from filteredInsertions to each Callable task
 			// calc backward dijkstra from pickup to ends of selected stops + starts
 			return toPickupPathSearch.calcPathDataMap(pickup, detourLinksSet.pickupDetourStartLinks.values(),
 					earliestPickupTime);
